+++ CHANGELOG.md
@@ -1,5 +1,48 @@
 
 
+# [8.12.0](https://github.com/zwave-js/zwave-js-ui/compare/v8.11.1...v8.12.0) (2023-04-03)
+
+
+### Bug Fixes
+
+* allow to call zwave apis in boot loader mode ([a286a8f](https://github.com/zwave-js/zwave-js-ui/commit/a286a8f399e3cc52948f0afa31f7a976f6fdf12f)), closes [#3023](https://github.com/zwave-js/zwave-js-ui/issues/3023)
+* remove driver ready check from `firmwareUpdateOTW` ([a793dd0](https://github.com/zwave-js/zwave-js-ui/commit/a793dd02a051459b99a186173e1dc07aa66e2b73))
+* **ui:** better default node name in compact view ([71a9884](https://github.com/zwave-js/zwave-js-ui/commit/71a988421679dc7e61ca5959f5a661b6f6c48c53))
+* **ui:** catch value write errors ([5195433](https://github.com/zwave-js/zwave-js-ui/commit/5195433469022574c810a33e907aa7978f6bf055)), closes [#3025](https://github.com/zwave-js/zwave-js-ui/issues/3025)
+* **ui:** hide manage nodes when there are selected ones ([be89889](https://github.com/zwave-js/zwave-js-ui/commit/be89889b7ef49e310aa2ff4dd03b145e3ef0d030))
+* **ui:** otw update not working in bootloader only mode ([89bb3e2](https://github.com/zwave-js/zwave-js-ui/commit/89bb3e2d3859d383f656a35eef490b81693e91ba)), closes [#3023](https://github.com/zwave-js/zwave-js-ui/issues/3023)
+
+
+### Features
+
+* **ui:** allow to send actions to selected nodes in table ([9806539](https://github.com/zwave-js/zwave-js-ui/commit/980653973e29e40f17f5b248da60693bb37e234a))
+* use valueId stored timestamp in MQTT JSON payload ([7b373c3](https://github.com/zwave-js/zwave-js-ui/commit/7b373c3d5956e4b46e1ffa762566fe04546e7a47)), closes [#3029](https://github.com/zwave-js/zwave-js-ui/issues/3029)
+
+## [8.11.1](https://github.com/zwave-js/zwave-js-ui/compare/v8.11.0...v8.11.1) (2023-03-24)
+
+
+### Features
+
+* bump zwave-js-server@1.27.0 ([#3018](https://github.com/zwave-js/zwave-js-ui/issues/3018)) ([0349aed](https://github.com/zwave-js/zwave-js-ui/commit/0349aedcc967a74ee3ea1f6a1b0f322b10ff7946))
+
+# [8.11.0](https://github.com/zwave-js/zwave-js-ui/compare/v8.10.1...v8.11.0) (2023-03-17)
+
+
+### Bug Fixes
+
+* **ui:** correctly print driver function result ([d6f42d7](https://github.com/zwave-js/zwave-js-ui/commit/d6f42d75aaf1d377fe52e13d9ec8c0f9b7b5ea93))
+
+
+### Features
+
+* ability to force query associations ([5154a02](https://github.com/zwave-js/zwave-js-ui/commit/5154a0276e41ae7b134a5a0db0b84df93327a00b)), closes [#2752](https://github.com/zwave-js/zwave-js-ui/issues/2752)
+* add `shutdownZwaveAPI` command ([6755044](https://github.com/zwave-js/zwave-js-ui/commit/67550449c30190f698c433172e9ba23b923aadcd))
+* add access-store-dir snippet ([83607e0](https://github.com/zwave-js/zwave-js-ui/commit/83607e0c6fc35af5248f4076ef3ab4bebc372da0))
+* allow to upload a file in store ([c00dfcc](https://github.com/zwave-js/zwave-js-ui/commit/c00dfccbf76253d51fd29db9ad93a47944b7d48f))
+* bump zwave-js@10.12.0 ([adc9dcc](https://github.com/zwave-js/zwave-js-ui/commit/adc9dcc8999a79b20ab12216d41ed53b800f53ad))
+* **ui:** better scan of small qr codes ([#3007](https://github.com/zwave-js/zwave-js-ui/issues/3007) ([295bec0](https://github.com/zwave-js/zwave-js-ui/commit/295bec0c50d20697812036e40e662b281ec27f68)), closes [#2767](https://github.com/zwave-js/zwave-js-ui/issues/2767)
+* **ui:** show low level security reason when node added with low security ([245f596](https://github.com/zwave-js/zwave-js-ui/commit/245f5963e27960584ec7c62c9ea1bfc918309ed0))
+
 ## [8.10.1](https://github.com/zwave-js/zwave-js-ui/compare/v8.10.0...v8.10.1) (2023-03-09)